+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add clj-toolbox.shelltools
 * Rename clj-toolbox.string-tools to clj-toolbox.strings
 * Add clj-toolbox.strings/split-whitespace
+* Add clj-toolbox.urls
 
 ## [0.9.0] — 2024-12-30
 * Add property tests using clojure.test.check for streams and base64